Add files via upload
This is a Blog that allow users to either register as an ordinary user or as a Moderator. A moderator have the right to publish a post and accept a like and comments from an ordinary user. The moderator can edit his/her post and profile. 
Upcoming Changes is to allow moderator to manage comments by deleting an unwanted comments.